moved components into appropriate folders , updated links, and reorganized
