+++ src/components/artists/ArtistList.vue
